requirements.txt
Flask==0.8
Flask-Login==0.1
Jinja2==2.6
Werkzeug==0.8.2
distribute==0.6.24
gunicorn==0.14.1
mongoengine==0.5.2
py-bcrypt==0.2
pymongo==2.1.1
slugify==0.0.1
wsgiref==0.1.2
